Put buffer protocol behind feature flag
