+++ pyo3-arrow/src/ffi/from_python/array.rs
@@ -1,6 +1,7 @@
 use crate::array::*;
+#[cfg(feature = "buffer_protocol")]
+use crate::buffer::AnyBufferProtocol;
 use crate::ffi::from_python::utils::call_arrow_c_array;
-use crate::input::AnyBufferProtocol;
 use pyo3::exceptions::PyValueError;
 use pyo3::prelude::*;
 use pyo3::{PyAny, PyResult};
@@ -10,9 +11,12 @@ impl<'a> FromPyObject<'a> for PyArray {
         if ob.hasattr("__arrow_c_array__")? {
             let (schema_capsule, array_capsule) = call_arrow_c_array(ob)?;
             Self::from_arrow_pycapsule(&schema_capsule, &array_capsule)
-        } else if let Ok(buf) = ob.extract::<AnyBufferProtocol>() {
-            Ok(buf.try_into()?)
         } else {
+            #[cfg(feature = "buffer_protocol")]
+            if let Ok(buf) = ob.extract::<AnyBufferProtocol>() {
+                return Ok(buf.try_into()?);
+            }
+
             Err(PyValueError::new_err(
                 "Expected object with __arrow_c_array__ method or implementing buffer protocol.",
             ))
